pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "s2and"
version = "0.1"
description = "Modified S2AND approach for author name disambiguation"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-core==2.3.3",
    "google-cloud-storage==2.10.0",
    "google-genai==1.35.0",
    "tenacity==9.1.2",
    "numpy==1.26.4",
    "scikit-learn==1.5.2",
    "text-unidecode==1.3",
    "requests==2.31.0",
    "hyperopt==0.2.4",
    "pandas==1.3.5",
    "lightgbm==4.5.0",
    "fastcluster==1.2.6",
    "pycld2==0.41",
    "fasttext-wheel==0.9.2",
    "shap==0.44.0",
    "matplotlib==3.7.0",
    "onnx==1.16.0",
    "onnxconverter-common==1.14.0",
    "onnxmltools==1.12.0",
    "onnxruntime==1.20.0",
    "skl2onnx==1.17.0",
    "seaborn==0.13.0",
    "torch==1.13.0",
    "tqdm==4.49.0",
    "strsimpy==0.2.0",
    "jellyfish==0.8.2",
    "PyYAML==6.0",
    "python-Levenshtein==0.12.2",
    "mlflow==2.20.0",
    "transformers==4.57.1",
    "ipykernel==6.16.2",
    "pyspark==3.3.1",
    "jarowinkler==1.2.3",
    "graphviz==0.20.1",
]

[project.urls]
Homepage = "https://github.com/atypon/S2AND.git"

[tool.setuptools.packages.find]
include = ["s2and"]